requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
black==21.12b0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
disnake==2.3.0
firebase-admin==5.2.0
frozenlist==1.2.0
google-api-core==2.4.0
google-api-python-client==2.35.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
msgpack==1.0.3
multidict==5.2.0
mypy-extensions==0.4.3
nose==1.3.7
packaging==21.3
pathspec==0.9.0
pkg_resources==0.0.0
platformdirs==2.4.1
proto-plus==1.19.8
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dotenv==0.19.2
python-randomword-fr==1.0.0
Random-Word==1.0.7
requests==2.27.1
rsa==4.8
six==1.16.0
tomli==1.2.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
yarl==1.7.2
